requirements.txt
python-telegram-bot==20.8
openai==1.25.2
telethon==1.29.3
apscheduler
nest_asyncio